requirements.txt
openvino==2023.3.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
pygame>=2.5.0
PyYAML>=6.0
